Remove promenade/helper
And make the same functionality available on the top level `Promenade`
module.

After some feedback from @balvig I decided that it is quite undesirable
for promenade to pollute the namespace of anything you want to
instrument.

This is simpler and looks just as nice - Indeed is probably better
because it's simple to understand which code is related to `Promenade`!